Moved functions from mcs module to package init. Deleted mcs module
